Moved Geotrellis DataSource into separate module, to be augmented with other DataSource-s.
