+++ datasource/src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,0 +1,1 @@
+astraea.spark.rasterframes.datasource.GeoTrellisDataSource
