--- src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,1 +1,0 @@
-astraea.spark.rasterframes.sources.GeoTrellisDataSource
